+++ src/main/java/com/Ultra_Nerd/CodeLyokoRemake15/init/LyokoArmorMaterial.java
@@ -16,35 +16,35 @@ public enum  LyokoArmorMaterial implements IArmorMaterial {
 
 
 
-     WARRIOR(Base.MOD_ID + ":william",33 , new int[]{11,14,16,11}, 420 , ModSounds.LAZY_WARRIOR.get(), 10F, () -> {
+     WARRIOR(Base.MOD_ID + ":william",33 , new byte[]{11,14,16,11}, 420 , ModSounds.LAZY_WARRIOR.get(), 10F, () -> {
         return Ingredient.fromItems(ModItems.PETABYTE.get());
     }),
-    BLANKHELM(Base.MOD_ID + ":blank",5 , new int[]{8,9,10,8}, 0 , SoundEvents.ITEM_ARMOR_EQUIP_CHAIN, 10F, () -> {
+    BLANKHELM(Base.MOD_ID + ":blank",5 , new byte[]{8,9,10,8}, 0 , SoundEvents.ITEM_ARMOR_EQUIP_CHAIN, 10F, () -> {
         return Ingredient.fromItems(ModItems.PETABYTE.get());
     }),
-    FELNINE(Base.MOD_ID + ":odd",25 , new int[]{4,7,9,4}, 420 , ModSounds.LAZY_FELINE.get(), 0F, () -> {
+    FELNINE(Base.MOD_ID + ":odd",25 , new byte[]{4,7,9,4}, 420 , ModSounds.LAZY_FELINE.get(), 0F, () -> {
         return Ingredient.fromItems(ModItems.GIGABYTE.get());
     }),
-    NINJA(Base.MOD_ID + ":yumi",24 , new int[]{1,4,7,2}, 420 , ModSounds.LAZY_NINJA.get(), 0F, () -> {
+    NINJA(Base.MOD_ID + ":yumi",24 , new byte[]{1,4,7,2}, 420 , ModSounds.LAZY_NINJA.get(), 0F, () -> {
         return Ingredient.fromItems(ModItems.MEGABYTE.get());
     }),
-    SAMURAI(Base.MOD_ID + ":ulrich",30 , new int[]{3,6,8,3}, 420 , ModSounds.LAZY_SAMURAI.get(), 10F, () -> {
+    SAMURAI(Base.MOD_ID + ":ulrich",30 , new byte[]{3,6,8,3}, 420 , ModSounds.LAZY_SAMURAI.get(), 10F, () -> {
         return Ingredient.fromItems(ModItems.GIGABYTE.get());
     }),
-    GUARDIAN(Base.MOD_ID + ":aelita",23 , new int[]{5,8,10,5}, 420 , ModSounds.LAZY_GUARDIAN.get(), 10F, () -> {
+    GUARDIAN(Base.MOD_ID + ":aelita",23 , new byte[]{5,8,10,5}, 420 , ModSounds.LAZY_GUARDIAN.get(), 10F, () -> {
         return Ingredient.fromItems(ModItems.TERABYTE.get());
     });
 
-    private static final int[] MAX_DAMAGE_ARRAY = new int[] {16,16,16,16};
+    private static final byte[] MAX_DAMAGE_ARRAY = new byte[] {16,16,16,16};
     private final String name;
     private final int maxDamageFactor;
-    private  final int[] damageReduction;
+    private  final byte[] damageReduction;
     private final int enchant;
     private final SoundEvent soundEvent;
     private final float toughness;
     private final LazyValue<Ingredient> redigitizeMaterial;
 
-    private LyokoArmorMaterial(String nameIn, int maxDamageFactor, int[] damageReduction, int enchant, SoundEvent soundsIn,
+    private LyokoArmorMaterial(String nameIn, int maxDamageFactor, byte[] damageReduction, int enchant, SoundEvent soundsIn,
                                float toughnessIn, Supplier<Ingredient> redigitizeMaterialIn)
     {
         this.name = nameIn;
